Restructure a project a little
