+++ src/logisim_converter/circuit/component.rs
