+++ src/logisim_converter/circuit/wire.rs
