+++ src/test_converter.rs
@@ -1,0 +1,5 @@
+use crate::logisim_converter::parse_logisim;
+
+pub fn test_converter() {
+    println!("{:?}", parse_logisim());
+}
